+++ test/blackbox-tests/test-cases/actions/stray-process.t
@@ -29,9 +29,9 @@ Now we stop Dune, which should normally kill all sub-processes:
   waiting for inotify sync
   waited for inotify sync
 
-  $ if kill -s 0 $CHILD_PID; then
+  $ if kill -s 0 $CHILD_PID 2> /dev/null; then
   >   echo "FAILURE: child is still running"
   > else
   >   echo "SUCCESS: child has exited"
   > fi
-  FAILURE: child is still running
+  SUCCESS: child has exited